requirements.txt
python-dotenv==0.19.0
SQLAlchemy==1.4.5
requests
beautifulsoup4
pandas